requirements.txt
flask
requests
regex
markdown
google-generativeai>=0.3.0
gunicorn
pymongo
werkzeug
python-dotenv
